requirements.txt
pyaudio
pocketsphinx
paho-mqtt
